Use isize for inputs and paramaters.

+++ src/lib.rs
@@ -1,7 +1,7 @@
 pub mod fib {
     use std::mem::replace;
 
-    fn validate_input(nth: i64) -> Result<bool, String> {
+    fn validate_input(nth: isize) -> Result<bool, String> {
         if nth >= 0 && nth < 186 {
             Ok(true)
         } else if nth > 185 {
@@ -13,7 +13,7 @@ pub mod fib {
 
     /// Calculate n'th Fibonacci number
     #[allow(dead_code)]
-    pub fn nth_fibonacci(nth: i64) -> Result<u128, String> {
+    pub fn nth_fibonacci(nth: isize) -> Result<u128, String> {
         validate_input(nth)?;
 
         let mut f0: u128 = 0;
@@ -28,7 +28,7 @@ pub mod fib {
 
     /// Return a vector of Fibonacci numbers up to n'th number
     #[allow(dead_code)]
-    pub fn fibonacci_to_nth(nth: i64) -> Result<Vec<u128>, String> {
+    pub fn fibonacci_to_nth(nth: isize) -> Result<Vec<u128>, String> {
         validate_input(nth)?;
 
         let mut v: Vec<u128> = Vec::new();
